+++ lovely/esdb/tests.py
@@ -85,6 +85,7 @@ def create_suite(testfile,
 
 def test_suite():
     return unittest.TestSuite((
+        create_suite('document/property.rst'),
         create_suite('document/document.rst'),
         create_suite('document/bulk.rst'),
     ))